replace existing runCmd function
