--- ui/tests/helpers/ui-console/selectors.ts
@@ -1,10 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: BUSL-1.1
- */
-
-export const REPL = {
-  toggle: '[data-test-console-toggle]',
-  consoleInput: '[data-test-component="console/command-input"] input',
-  logOutputItems: '[data-test-component="console/output-log"] > div',
-};
